Generalize the evaluation of defined and undefined
`expr is undefined` evaluates to true if `expr` evaluates to Tnull and the
`defined` test does the opposite.

This is closer to what Jinja2 does with its Undefined class and should enable
the use of these tests with any value.

Caveat: Tnull is considered to be undefined although None is seen as defined by
Jinja2.